+++ PayPal/src/main/java/com/braintreepayments/api/paypal/PayPalRequest.kt
@@ -1,6 +1,5 @@
 package com.braintreepayments.api.paypal
 
-import android.net.Uri
 import android.os.Parcelable
 import androidx.annotation.RestrictTo
 import com.braintreepayments.api.core.Authorization
@@ -74,9 +73,6 @@ import org.json.JSONException
  * @property userPhoneNumber User phone number used to initiate a quicker authentication flow in
  * cases where the user has a PayPal Account with the phone number.
  * @property lineItems The line items for this transaction. It can include up to 249 line items.
- * @property shippingCallbackUrl Server side shipping callback URL to be notified when a customer
- * updates their shipping address or options. A callback request will be sent to the merchant server
- * at this URL.
  */
 abstract class PayPalRequest internal constructor(
     open val hasUserLocationConsent: Boolean,
@@ -91,8 +87,7 @@ abstract class PayPalRequest internal constructor(
     open var riskCorrelationId: String? = null,
     open var userAuthenticationEmail: String? = null,
     open var userPhoneNumber: PayPalPhoneNumber? = null,
-    open var lineItems: List<PayPalLineItem> = emptyList(),
-    open var shippingCallbackUrl: Uri? = null
+    open var lineItems: List<PayPalLineItem> = emptyList()
 ) : Parcelable {
 
     @Throws(JSONException::class)