+++ PayPal/src/test/java/com/braintreepayments/api/paypal/PayPalVaultRequestUnitTest.java
@@ -7,7 +7,6 @@ import static junit.framework.Assert.assertTrue;
 import static org.junit.Assert.assertSame;
 import static org.mockito.Mockito.mock;
 
-import android.net.Uri;
 import android.os.Build;
 import android.os.Parcel;
 
@@ -17,7 +16,6 @@ import com.braintreepayments.api.core.PostalAddress;
 import com.braintreepayments.api.testutils.Fixtures;
 
 import org.json.JSONException;
-import org.json.JSONObject;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
@@ -300,58 +298,5 @@ public class PayPalVaultRequestUnitTest {
         );
 
         assertTrue(requestBody.contains("\"phone_number\":{\"country_code\":\"1\",\"national_number\":\"1231231234\"}"));
-    }
-
-    @Test
-    public void createRequestBody_sets_shippingCallbackUri_when_not_null() throws JSONException {
-        String urlString = "https://www.example.com/path";
-        Uri uri = Uri.parse(urlString);
-
-        PayPalVaultRequest request = new PayPalVaultRequest(true);
-        request.setShippingCallbackUrl(uri);
-
-        String requestBody = request.createRequestBody(
-                mock(Configuration.class),
-                mock(Authorization.class),
-                "success_url",
-                "cancel_url",
-                null
-        );
-
-        JSONObject jsonObject = new JSONObject(requestBody);
-        assertEquals(urlString, jsonObject.getString("shipping_callback_url"));
-    }
-
-    @Test
-    public void createRequestBody_does_not_set_shippingCallbackUri_when_null() throws JSONException {
-        PayPalVaultRequest request = new PayPalVaultRequest(true);
-
-        String requestBody = request.createRequestBody(
-                mock(Configuration.class),
-                mock(Authorization.class),
-                "success_url",
-                "cancel_url",
-                null
-        );
-
-        JSONObject jsonObject = new JSONObject(requestBody);
-        assertFalse(jsonObject.has("shipping_callback_url"));
-    }
-
-    @Test
-    public void createRequestBody_does_not_set_shippingCallbackUri_when_empty() throws JSONException {
-        PayPalVaultRequest request = new PayPalVaultRequest(true);
-        request.setShippingCallbackUrl(Uri.parse(""));
-
-        String requestBody = request.createRequestBody(
-                mock(Configuration.class),
-                mock(Authorization.class),
-                "success_url",
-                "cancel_url",
-                null
-        );
-
-        JSONObject jsonObject = new JSONObject(requestBody);
-        assertFalse(jsonObject.has("shipping_callback_url"));
     }
 }